+++ catalyst/metrics/_auc.py
@@ -1,9 +1,10 @@
-from typing import Dict
+from typing import Dict, Tuple
 
 import torch
 
 from catalyst.metrics._metric import ICallbackLoaderMetric
-from catalyst.metrics.functional._auc import auc
+from catalyst.metrics.functional._auc import auc, binary_auc
+from catalyst.metrics.functional._misc import process_multilabel_components
 from catalyst.utils.distributed import all_gather, get_rank
 
 
@@ -14,6 +15,10 @@ class AUCMetric(ICallbackLoaderMetric):
         compute_on_call: if True, computes and returns metric value during metric call
         prefix: metric prefix
         suffix: metric suffix
+
+    .. warning::
+
+        This metric is under API improvement.
     """
 
     def __init__(self, compute_on_call: bool = True, prefix: str = None, suffix: str = None):
@@ -22,7 +27,7 @@ class AUCMetric(ICallbackLoaderMetric):
         self.metric_name = f"{self.prefix}auc{self.suffix}"
         self.scores = []
         self.targets = []
-        self._is_ddp = False
+        self._is_ddp = get_rank() > -1
 
     def reset(self, num_batches, num_samples) -> None:
         """Resets all fields"""
@@ -40,7 +45,7 @@ class AUCMetric(ICallbackLoaderMetric):
         self.scores.append(scores.cpu().detach())
         self.targets.append(targets.cpu().detach())
 
-    def compute(self) -> torch.Tensor:
+    def compute(self) -> Tuple[torch.Tensor, float, float, float]:
         """Computes the AUC metric based on saved statistics."""
         targets = torch.cat(self.targets)
         scores = torch.cat(self.scores)
@@ -50,19 +55,25 @@ class AUCMetric(ICallbackLoaderMetric):
             scores = torch.cat(all_gather(scores))
             targets = torch.cat(all_gather(targets))
 
-        score = auc(outputs=scores, targets=targets)
-        return score
+        scores, targets, _ = process_multilabel_components(outputs=scores, targets=targets)
+        per_class = auc(scores=scores, targets=targets)
+        micro = binary_auc(scores=scores.view(-1), targets=targets.view(-1))[0]
+        macro = per_class.mean().item()
+        weights = targets.sum(axis=0) / len(targets)
+        weighted = (per_class * weights).sum().item()
+        return per_class, micro, macro, weighted
 
     def compute_key_value(self) -> Dict[str, float]:
         """Computes the AUC metric based on saved statistics and returns key-value results."""
-        per_class_auc = self.compute()
+        per_class_auc, micro_auc, macro_auc, weighted_auc = self.compute()
         output = {
             f"{self.metric_name}/class_{i:02d}": value.item()
             for i, value in enumerate(per_class_auc)
         }
-        macro_auc = per_class_auc.mean().item()
+        output[f"{self.metric_name}/_micro"] = micro_auc
         output[self.metric_name] = macro_auc
         output[f"{self.metric_name}/_macro"] = macro_auc
+        output[f"{self.metric_name}/_weighted"] = weighted_auc
         return output